+++ NFixtureFactoryTest/Model/Client.cs
@@ -7,7 +7,7 @@ namespace NFixtureFactoryTest.Model
 	{
 		public String Name {get;set;}
 		public Address Address { get; set; }
-		public Double CPF {get;set;}
+		public Decimal CPF {get;set;}
 		public IList<Phone> Phones { get; set;}
 	}
 }
